Rename reducer files
