+++ packages/noya-state/src/index.ts
@@ -3,10 +3,10 @@ export * as Selectors from './selectors/selectors';
 export * as Models from './models';
 export * as TextStyleSelectors from './selectors/textStyleSelectors';
 
-export * from './reducers/workspace';
-export * from './reducers/application';
-export * from './reducers/history';
-export * from './reducers/interaction';
+export * from './reducers/workspaceReducer';
+export * from './reducers/applicationReducer';
+export * from './reducers/historyReducer';
+export * from './reducers/interactionReducer';
 export * from './types';
 
 export type { SelectionType } from './utils/selection';
