+++ packages/noya-state/src/selectors/geometrySelectors.ts
@@ -14,8 +14,8 @@ import * as Primitives from 'noya-renderer/src/primitives';
 import { EnterReturnValue, SKIP, STOP } from 'tree-visit';
 import { ApplicationState, Layers, PageLayer } from '../index';
 import { visitReversed } from '../layers';
-import { CompassDirection } from '../reducers/interaction';
-import { CanvasInsets } from '../reducers/workspace';
+import { CompassDirection } from '../reducers/interactionReducer';
+import { CanvasInsets } from '../reducers/workspaceReducer';
 import type { Point, Rect } from '../types';
 import { getSelectedLayerIndexPaths } from './indexPathSelectors';
 import { getCurrentPage } from './pageSelectors';
